refactor(wallet): rename InMemoryKey.ts to match type
