+++ packages/wallet/src/KeyManagement/index.ts
@@ -1,4 +1,4 @@
 export * as errors from './errors';
 export * from './KeyManager';
-export * from './InMemoryKey';
+export * from './InMemoryKeyManager';
 export * as util from './util';
